Consistently call before and after callbacks.
Calling of the callback_before_*() and callback_after_*() functions was
not done consistently in the XML and PBF input classes. Now they are
always called before and after each run of several objects of the same
type, respectively. This will do the right thing if the input file has
all nodes, ways, and relations together but is independent of the ordering
between object types.