Remove callback_object()
The object callback was always a bit odd. It was called for all objects (ie
nodes, ways, and relations) before the type-specific object. Some handlers
used it for things that need to be done regardless of object type. But it
was limited in what it could do, because it was always called before the
other callback. What if you wanted common things done after the other stuff?
Or in the middle? Or only for nodes and ways, but not for relations?

Its more straightforward to decide in your callback code if there is some
commonality and facter that out into another method in your handler class.

Also it was unclear whether it should be called for multipolygons (it wasn't).

So I decided to remove it.