Refactoring of timestamp as string output.
The method get_timestamp_str() returning a char* to a static buffer
was retired in favour of get_timestamp_as_string() which return
a std::string and is thread safe.

Osmium::OSM::Object::max_length_timestamp is not available any more.
Its not actually needed by any code outside the lib itself and it
had the wrong name anyway (its not the max length, the timestamp
always has this length).