Great reorganisation of file handling.
There is a new class Osmium::OSMFile through which all I/O of OSM files is
handled. It can work with PBF and (compressed or uncompressed) XML files and
even transparently read from the network with the help of curl.

There are still a few thing to do, but a flexible infrastructure is there.