Refactoring of way node list.
There is a new class WayNodeList which contains a list of WayNodes, which in
turn contain an ID and a Position. The WayNodeList class is now used in ways.
The idea is to make this a bit more flexible in the long run.
This commit also introduces some unit tests using the Boost Unit Test framework.