Use shared_ptr instead of simple * in handlers
The node(), way(), and relation() methods in handlers used to take simple
pointers to the objects. Now they take references to const shared_ptr's
instead. Together with changes to the input classes, this allows handlers to
keep the objects without a costly copy. Most handlers don't need this, but
there are some cases.

This change does not lead to much extra overhead (I benchmarked it), but it
makes some use cases much easier in the handlers. The first use case is in
the new EndTime handler: All objects now have an endtime in addition to the
timestamp. When reading an OSM history file, the EndTime handler will add
the endtime to all but the last version of an object. That way it is easier
to see the time range in which an object version was valid.