New ObjectStore can store OSM objects.
It can be filled by using it as a handler. It can feed its contents to another
handler. And there is a special handler that merges its input with the content
of the ObjectStore. There is an example program osmium_store_and_debug to show
how it works.
The ObjectStore can be used to apply diffs (for instance from .osc files) if
the contents of those files fits into memory.