Turned runtime optimization into compiletime.
The methods in Osmium::Input::PBF that parse blocks of nodes, ways, and
relations will not be called when there is no handler waiting for the data (or,
to be more correct, when the handler is the empty handler Osmium::Handler::Base).
We did this with a runtime check using typeid, now its done with a compiletime
check using overloading tricks.