Add support for some automagic configuration.
The new script get_options.sh can be called to find out which compiler and link
options must be used to compile any Osmium header file or C++ file including
those headers. This is done by defining macros of the form
OSMIUM_COMPILE_WITH_CFLAGS_* and OSMIUM_COMPILE_WITH_LIBS_* in the header files.
The check-includes target in the Makefile uses this to figure out how to do the
compile.