Put OGR transaction calls into osmium_toogr.
Using transactions makes writing to SQLite files significantly faster. But it
seems to make writing to Shapefiles a bit slower. When the files get really
large it might be better to commit occasionally in between. Your mileage may
vary. Somebody has to properly benchmark this at some point...

So for the time being those calls have been put in but commented out just to
give everybody an idea that they could be used.