Changed the way debugging is done.
This has changed several times now over the life of Osmium. I hope it is
settling now... Debug support is still controlled by the OSMIUM_WITH_DEBUG
macro, but there is only one place in the code where it is used to set a const
thats then used everywhere else. This should make the code more robust and
easier to read while still disabling debug code at compile time if it is
not needed.