Make osmium_time work on Windows.
osmium_time outputs wallclock time and user/system time, the latter is not
available on Windows. With this change only wallclock time is reported which is
the most important anyway. On unix systems you can set
REPORT_USER_AND_SYSTEM_TIME when compiling and it will also report the other
values.