Split up Osmium::Ser::Handler
The basic functionality is still in Handler, but the update of maps from
member ids to ids of objects moved to the new UpdateHandler. The
UpdateHandler gets a link to the old version of an object from the
Handler (if there is an old version and handler runs in update_mode).
Currently it doesn't use the old version, that code still needs to be
added.