Hacked together hybrid map
The Member::Vector class is much more space efficient the Member::MultiMap. But
adding a member at the end destroys the sorting. So this Hybrid class wraps a
Vector and a MultiMap class in a way that entries are added the MultiMap and
then consolidated into the Vector occasionally. Well, the current implementation
is very hacky. Just to see where this goes.